Move standaloneTaskSid ou of ContactState
